modified

+++ src/main.h
@@ -5,7 +5,7 @@
 
 void calTotalField(traps alltraps[],double *totalField, int totSteps);
 void cleanTrapsField(traps tp[]);
-double findFieldMin_V2(double *tField, int rowNum);
+int findFieldMin_V2(double *tField, int rowNum);
 //double findFieldMax(double tField[]);
 
 #endif
